Refactor networking to use streaming/RESP parser and improve logging

+++ lib/ThreadPool.h
@@ -11,30 +11,34 @@
 
 class ThreadPool;
 
-class Worker {
-  ThreadPool *_pool;
+class Worker
+{
+    ThreadPool* _pool;
 
 public:
-  Worker(ThreadPool *pool) : _pool(pool) {}
-  void operator()();
+    Worker(ThreadPool* pool) : _pool(pool)
+    {
+    }
+    void operator()();
 };
 
-class ThreadPool {
-  friend class Worker;
-  bool _stop;
-  std::mutex _queueMutex;
-  std::condition_variable _cv;
+class ThreadPool
+{
+    friend class Worker;
+    bool _stop;
+    std::mutex _queueMutex;
+    std::condition_variable _cv;
 
-  std::vector<std::thread> _workers;
-  std::queue<std::function<void()>> _tasks;
+    std::vector<std::thread> _workers;
+    std::queue<std::function<void()>> _tasks;
 
 public:
-  ThreadPool(int n = std::thread::hardware_concurrency());
-  void Enqueue(std::function<void()> task);
-  ~ThreadPool();
+    ThreadPool(size_t n = std::thread::hardware_concurrency());
+    void Enqueue(std::function<void()> task);
+    ~ThreadPool();
 
 private:
-  std::function<void()> Dequeue();
-  ThreadPool(const ThreadPool &) = delete;
-  ThreadPool &operator=(const ThreadPool &) = delete;
+    std::function<void()> Dequeue();
+    ThreadPool(const ThreadPool&) = delete;
+    ThreadPool& operator=(const ThreadPool&) = delete;
 };
